Reorganized scripts and removed old dependencies
- vsphere scripts are now located in vsphere/
- vsphere scripts are now their own CLI command. They can be invoked using "vsphere", or "python -m adles.vsphere". This will be the template used for future scripts for other platforms.
- Removed docopt and argopt from dependencies.
- Removed old entrypoints.
- Reorganized utils.
