+++ adles/scripts/script_utils.py
@@ -11,37 +11,8 @@
 # limitations under the License.
 
 import logging
-import sys
 
-from adles.utils import read_json, handle_keyboard_interrupt
-
-
-# TODO: cli args instead of filename
-# e.g. "--vsphere-user" on all vsphere scripts, so ConfigArgParse can do it's thing globally
-def make_vsphere(filename=None):
-    """
-    Creates a vSphere object using either a JSON file or by prompting the user.
-
-    :param str filename: Name of JSON file with connection info
-    :return: vSphere object
-    :rtype: :class:`Vsphere`
-    """
-    from adles.vsphere.vsphere_class import Vsphere
-    if filename is not None:
-        info = read_json(filename)
-        if info is None:
-            sys.exit(1)
-        return Vsphere(username=info.get("user"),
-                       password=info.get("pass"),
-                       hostname=info.get("host"),
-                       port=info.get("port", 443),
-                       datacenter=info.get("datacenter"),
-                       datastore=info.get("datastore"))
-    else:
-        logging.info("Enter information to connect to the vSphere environment")
-        datacenter = input("Datacenter  : ")
-        datastore = input("Datastore   : ")
-        return Vsphere(datacenter=datacenter, datastore=datastore)
+from adles.utils import handle_keyboard_interrupt
 
 
 @handle_keyboard_interrupt
@@ -114,30 +85,3 @@ def user_input(prompt, obj_name, func):
                   % (obj_name, item_name))
 
 
-def resolve_path(server, thing, prompt=""):
-    """
-    This is a hacked together script utility to get folders or VMs.
-
-    :param server: Vsphere instance
-    :type server: :class:`Vsphere`
-    :param str thing: String name of thing to get (folder | vm)
-    :param str prompt: Message to display
-    :return: (thing, thing name)
-    :rtype: tuple(vimtype, str)
-    """
-    from adles.vsphere.vm import VM
-    if thing.lower() == "vm":
-        get = server.get_vm
-    elif thing.lower() == "folder":
-        get = server.get_folder
-    else:
-        logging.error("Invalid thing passed to resolve_path: %s", thing)
-        raise ValueError
-
-    res = user_input("Name of or path to %s %s: " % (thing, prompt), thing,
-                     lambda x: server.find_by_inv_path("vm/" + x)
-                     if '/' in x else get(x))
-    if thing.lower() == "vm":
-        return VM(vm=res[0]), res[1]
-    else:
-        return res
